refactor: remove dead generic-JSON persistence plumbing
Phase 2 step 6. With project/dataset/annotation on per-module Diesel, the
Phase-1 generic JSON layer has no users:
- Deleted src/domain/common/ (CrudRepository/JsonCrudRepository + the
  Tauri-coupled save_entity/delete_entity free fns + the HasId alias).
- Deleted shared::EntitySource (the generic JSON port); moved PortError into
  shared::error (kept for EventPublisher); event.rs imports updated.
- Removed composition::EntitySourceAdapter; TauriEventPublisher + the
  From<DomainError> mapping remain.
The residual DesktopStore keeps its typed methods + the EntityStore trait for
the not-yet-migrated entities (annotations/history/settings/ai/...), which
AnalysisService and the lib.rs command helpers still use. Full cargo test green
(42 binary + 5 runtime-manager + 4 arch + 1 shared).
