--- apps/studio/src-tauri/src/domain/common/mod.rs
@@ -1,2 +1,0 @@
-pub mod repository;
-pub mod service;
