+++ apps/studio/src-tauri/src/domain/mod.rs
@@ -1,7 +1,6 @@
 pub mod ai;
 pub mod analysis;
 pub mod cloud;
-pub mod common;
 pub mod images;
 pub mod labels;
 pub mod projects;
